HW3/flight_service/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.3
tenacity==8.2.3
